Move shared web socket models to shared-desktop library
